# Conflicts:
#	open-metadata-resources/open-metadata-demos/virtual-data-connector/virtualizer/src/test/resources/json/business.json
#	open-metadata-resources/open-metadata-demos/virtual-data-connector/virtualizer/src/test/resources/json/technical.json
#	open-metadata-resources/open-metadata-demos/virtualizer-deprecated/src/main/java/org/odpi/openmetadata/virtualdataconnector/virtualiser/views/ViewsConstructor.java

+++ open-metadata-implementation/adapters/open-connectors/repository-services-connectors/open-metadata-collection-store-connectors/ibm-igc-repository-connector/igc-repository-connector/src/main/java/org/odpi/openmetadata/adapters/repositoryservices/igc/repositoryconnector/mapping/relationships/DataClassAssignmentMapper.java
@@ -259,7 +259,7 @@ public class DataClassAssignmentMapper extends RelationshipMapping {
                                     repositoryName,
                                     relationshipProperties,
                                     "valueFrequency",
-                                    (Long) valFreq,
+                                    ((Double) valFreq).longValue(),
                                     methodName
                             );
                         }
@@ -445,7 +445,7 @@ public class DataClassAssignmentMapper extends RelationshipMapping {
                                     repositoryName,
                                     relationshipProperties,
                                     "valueFrequency",
-                                    (Long) valFreq,
+                                    ((Double) valFreq).longValue(),
                                     methodName
                             );
                         }
